+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import NavBar from './components/NavBar/NavBar';
 import { Main } from './containers/Main/Main';
-import Cart from './containers/Cart/Cart';
+import Cart from './components/Cart/Cart';
 import Description from './components/Description/Description';
 import { catalog } from './utilities/catalog';
 import { displayRandomMessage } from './utilities/utilities'
@@ -13,6 +13,7 @@ class App extends React.Component {
     this.state = {
       cart: [],
       toggleCart: false,
+      cartFullPage: false,
       showMessage: false,
       showDescription: false,
       currentDisplayObject: '',
@@ -20,8 +21,9 @@ class App extends React.Component {
       currentItems: catalog
 
     }
-    this.toggleCart = this.toggleCart.bind(this);
     this.addToCart = this.addToCart.bind(this);
+    this.toggleCart = this.toggleCart.bind(this);
+    this.fullViewCart = this.fullViewCart.bind(this);
     this.showDescription = this.showDescription.bind(this);
     this.closeDescription = this.closeDescription.bind(this);
     this.searchItem = this.searchItem.bind(this);
@@ -40,9 +42,34 @@ class App extends React.Component {
     })), 1500);
   }
 
+  adjustQuantity(key, target) {
+    if (key === 'add') {
+      return this.setState(state => ({
+        cart: [...state.cart, target]
+      }));
+    }
+    if (key === 'remove') {
+      let shallowCopy = Array.from(this.state.cart).filter((element, index, arr) => index !== arr.lastIndexOf(target))
+      return this.setState(state => ({
+        cart: shallowCopy
+      }))
+    }
+    if (key === 'delete') {
+      return this.setState(state => ({
+        cart: [...state.cart.filter(el => el !== target)]
+      }))
+    }
+  }
+
   toggleCart() {
     this.setState(state => ({
       toggleCart: !state.toggleCart
+    }))
+  }
+
+  fullViewCart() {
+    this.setState(state => ({
+      fullPage: !state.cartFullPage
     }))
   }
 
@@ -69,7 +96,14 @@ class App extends React.Component {
       <div className='wrapper'>
         <NavBar firstLink='Home' secondLink='About' onClick={this.toggleCart} handleSearch={this.searchItem} />
         <Main catalog={this.state.currentItems} click={this.addToCart} showDescription={this.showDescription} />
-        {this.state.toggleCart ? <Cart clickToggle={this.toggleCart}  showDescription={this.showDescription} cart={this.state.cart}/> : null}
+        {this.state.toggleCart ? <Cart 
+                                 clickToggle={this.toggleCart}  
+                                 showDescription={this.showDescription} 
+                                 cart={this.state.cart} 
+                                 adjustQuantity={this.adjustQuantity} 
+                                 fullPage={this.state.cartFullPage}
+                                 fullView={this.state.fullViewCart}
+                                 /> : null}
         {this.state.showMessage ? displayRandomMessage() : null}
         {this.state.showDescription ? <Description displayObject={this.state.currentDisplayObject} onClick={this.closeDescription} /> : null}
       </div>
